+++ apps/www/pages/permalinks/[hash]/index.tsx
@@ -1,7 +1,7 @@
 import { GetObjectCommandOutput, S3Client } from "@aws-sdk/client-s3"
 import { TimestampView } from "@phylopic/ui"
 import { Hash, isHash } from "@phylopic/utils"
-import { getJSON } from "@phylopic/utils-aws"
+import { createAwsClientConfig, getJSON } from "@phylopic/utils-aws"
 import type { GetStaticPaths, GetStaticProps, NextPage } from "next"
 import { NextSeo } from "next-seo"
 import PageLayout, { Props as PageLayoutProps } from "~/pages/PageLayout"
@@ -60,13 +60,13 @@ export const getStaticProps: GetStaticProps<Props, { hash: Hash }> = async conte
     if (!isHash(hash)) {
         return { notFound: true }
     }
-    const client = new S3Client({
-        credentials: {
-            accessKeyId: process.env.S3_ACCESS_KEY_ID ?? "",
-            secretAccessKey: process.env.S3_SECRET_ACCESS_KEY ?? "",
-        },
-        region: process.env.S3_REGION,
-    })
+    const client = new S3Client(
+        createAwsClientConfig({
+            region: process.env.S3_REGION!,
+            accessKeyIdEnv: "S3_ACCESS_KEY_ID",
+            secretAccessKeyEnv: "S3_SECRET_ACCESS_KEY",
+        }),
+    )
     let data: PermalinkData
     let output: GetObjectCommandOutput
     try {