+++ packages/utils-aws/src/credentials/index.ts
@@ -1,0 +1,2 @@
+export { default as createAwsClientConfig } from "./createAwsClientConfig"
+export type { CreateAwsClientConfigOptions } from "./createAwsClientConfig"