+++ packages/utils-aws/src/index.ts
@@ -1,2 +1,3 @@
+export * from "./credentials/index"
 export * from "./errors/index"
 export * from "./s3/index"